requirements.txt
tweepy
python-dotenv
groq
starlette
uvicorn